requirements.txt
openai==1.102.0
pillow==11.2.1
pillow-heif==1.3.0
python-dotenv==1.2.2
